% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gic_csgl.R
\name{calc.gic}
\alias{calc.gic}
\title{Calculates generalized information criterion}
\usage{
calc.gic(Z, y, betahat, int = 0)
}
\arguments{
\item{Z}{Matrix of log OTU proportions}

\item{y}{Outcome vector}

\item{betahat}{Estimated coefficient vector}

\item{int}{Estimated intercept; default = 0}
}
\value{
GIC
}
\description{
Calculates GIC as defined in Lin (2014).
}
