% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgl.R
\name{calc.lammax}
\alias{calc.lammax}
\title{Calculates maximum lambda value}
\usage{
calc.lammax(Z, y, n, p, H, mu, theta, groups, thresh, erel, maxit)
}
\arguments{
\item{Z}{Matrix of log OTU proportions}

\item{y}{Outcome vector}

\item{n}{Number of subjects}

\item{p}{Number of OTUs}

\item{H}{Matrix I_p - (1/p) 1_p 1_p'}

\item{mu}{Augmented Lagrangian parameter}

\item{theta}{Defines convex combination between L1 and L2 penalties
(theta = 1 is L1 only, theta = 0 is L2 only)}

\item{groups}{Vector indicating group membership of each OTU}

\item{thresh}{Threshold for convergence}

\item{erel}{Internal parameter for ADMM convergence}

\item{maxit}{Maximum number of iterations}
}
\value{
Maximum lambda value
}
\description{
Calculates maximum lambda value as smallest lambda for which
all coefficients are zero.
}
