% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgl_fit.R
\name{csgl.fit}
\alias{csgl.fit}
\title{Fits CSGL}
\usage{
csgl.fit(Z, y, n, p, H, mu, lambda, theta, groups, thresh, erel, maxit)
}
\arguments{
\item{Z}{Matrix of log OTU proportions}

\item{y}{Outcome vector}

\item{n}{Number of subjects}

\item{p}{Number of OTUs}

\item{H}{Matrix I_p - (1/p) 1_p 1_p'}

\item{mu}{Augmented Lagrangian parameter}

\item{lambda}{Regularization parameter}

\item{theta}{Defines convex combination between L1 and L2 penalties
(theta = 1 is L1 only, theta = 0 is L2 only)}

\item{groups}{Vector indicating group membership of each OTU}

\item{thresh}{Threshold for convergence}

\item{erel}{Internal parameter for ADMM convergence}

\item{maxit}{Maximum number of iterations}
}
\value{
Fitted coefficients beta
}
\description{
Function to fit CSGL. Interior function for csgl (sequence of lambdas),
gic.csgl (choosing best lambda by GIC), and cv.csgl (choosing best
lambda by k-fold cross-validation).
}
