% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgl.R
\name{csgl}
\alias{csgl}
\title{Compositional sparse group lasso}
\usage{
csgl(Z, y, groups, mu = 1, theta = 0.95, lam.seq = NULL, nlam = 25,
  min.frac = 0.1, thresh = 1e-07, maxit = 10000, std = T, verbose = F)
}
\arguments{
\item{Z}{Matrix of log OTU proportions}

\item{y}{Outcome vector}

\item{groups}{Vector indicating group membership of each OTU}

\item{mu}{Augmented Lagrangian parameter}

\item{theta}{Defines convex combination between L1 and L2 penalties
(theta = 1 is L1 only, theta = 0 is L2 only)}

\item{lam.seq}{Sequence of lambda values to consider}

\item{nlam}{Number of lambda values to try}

\item{min.frac}{Minimum value of the penalty parameter, as a fraction of the maximum value}

\item{thresh}{Threshold for convergence}

\item{maxit}{Maximum number of iterations}

\item{std}{Logical flag for variable standardization before fitting the model}

\item{verbose}{Logical flag for printing updates during model fitting}
}
\value{
List with components
\item{int}{Vector of intercepts, one per lambda}
\item{beta}{Matrix of coefficients, one column per lambda}
\item{lambdas}{Sequence of lambda values}
}
\description{
Fits CSGL for a sequence of lambda values.
}
