% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_csgl.R
\name{getFolds}
\alias{getFolds}
\title{Gets folds for cross-validation}
\usage{
getFolds(n, nfolds)
}
\arguments{
\item{n}{Number of subjects}

\item{nfolds}{Desired number of folds}
}
\value{
List of folds
}
\description{
Gets folds for cross-validation
}
