% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_updates.R
\name{update.alpha}
\alias{update.alpha}
\title{Update alpha vector}
\usage{
\method{update}{alpha}(H, beta, xi, p)
}
\arguments{
\item{H}{Matrix I_p - (1/p) 1_p 1_p'}

\item{beta}{p-vector of coefficients}

\item{xi}{p-vector of Lagrange multipliers}

\item{p}{Number of OTUs}
}
\value{
Updated alpha vector
}
\description{
This function updates alpha by projecting (beta + xi) onto the subspace
defined by sum(alpha) = 0.
}
