% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_updates.R
\name{update.beta}
\alias{update.beta}
\title{Update beta vector}
\usage{
\method{update}{beta}(y, Z, n, p, mu, alpha, xi, lambda, theta, groups)
}
\arguments{
\item{y}{Outcome vector}

\item{Z}{Matrix of log OTU proportions}

\item{n}{Number of subjects (nrow(Z))}

\item{p}{Number of OTUs (ncol(Z))}

\item{mu}{Augmented Lagrangian parameter}

\item{alpha}{p-vector of coefficients}

\item{xi}{p-vector of Lagrange multipliers}

\item{lambda}{Regularization parameter}

\item{theta}{Defines convex combination between L1 and L2 penalties (theta = 1 is L1 only)}

\item{groups}{Vector indicating group membership of each OTU}
}
\value{
Updated beta vector
}
\description{
This function uses the SGL package to minimize the augmented Lagrangian
with respect to beta.
}
