% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_updates.R
\name{update.xi}
\alias{update.xi}
\title{Update xi vector}
\usage{
\method{update}{xi}(alpha, beta, xi)
}
\arguments{
\item{alpha}{p-vector of coefficients}

\item{beta}{p-vector of coefficients}

\item{xi}{p-vector of Lagrange multipliers}
}
\value{
Updated xi vector
}
\description{
This function updates the Lagrange multipliers xi.
}
